platform/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "alembic*", "scripts*"]

[project]
name = "securesync-platform"
version = "0.1.0"
description = "SecureSync Platform API — orchestration layer for the NEXO AI security audit platform"
requires-python = ">=3.12"
authors = [{ name = "NEXO AI" }]

dependencies = [
    # Core
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.0",
    "pydantic==2.7.0",
    "pydantic-settings==2.3.0",
    "sqlalchemy[asyncio]==2.0.30",
    "alembic==1.13.1",
    # Async DB driver. asyncpg is faster but has unresolved Windows
    # ProactorEventLoop issues (WinError 64); psycopg3 works everywhere.
    "psycopg[binary,pool]==3.1.19",
    "celery==5.4.0",
    "redis==5.0.4",
    "apscheduler==3.10.4",
    "croniter==3.0.4",               # cron-expression parser for scan schedules

    # Security
    "python-jose[cryptography]==3.3.0",
    "passlib==1.7.4",
    "bcrypt==4.0.1",                  # passlib 1.7.4 ↔ bcrypt 5.x ABI break workaround
    "python-multipart==0.0.9",

    # HTTP
    "httpx==0.27.0",
    "dnspython==2.7.0",              # TXT lookup for custom-domain verification

    # Scanner integrations
    "python-gvm==24.1.0",
    "python-nmap==0.7.1",
    "python-owasp-zap-v2.4==0.0.21",

    # Reporting
    "weasyprint==62.1",
    "jinja2==3.1.4",

    # Utilities
    "structlog==24.2.0",
    "tenacity==8.3.0",
    "email-validator==2.1.1",

    # Observability
    "sentry-sdk[fastapi,celery]==2.14.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.4.4",
    "mypy==1.10.0",
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
    "asgi-lifespan==2.1.0",
    "factory-boy==3.3.0",
    "respx==0.23.1",         # httpx mocking for integration-client tests
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "ASYNC", "S", "RET", "SIM", "TCH"]
ignore = [
    "S101",  # assert allowed in tests
    "B008",  # FastAPI Depends() in defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/versions/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/main.py", "alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
